# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple CRUD Task API built with Express"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "description": "Returns a list of all tasks",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      },
      "post": {
        "summary": "Create a task",
        "description": "Creates a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "Learn Swagger"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },

    "/tasks/{id}": {
      "get": {
        "summary": "Get one task",
        "description": "Returns a task by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task found"
          },
          "404": {
            "description": "Task not found"
          }
        }
      },

      "put": {
        "summary": "Update a task",
        "description": "Updates title or completion status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "title": "Updated task",
                "done": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated"
          }
        }
      },

      "delete": {
        "summary": "Delete a task",
        "description": "Deletes a task by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted"
          }
        }
      }
    }
  }
}